+++ sql/updates/h3--0.3.2--0.4.0.sql
@@ -17,7 +17,7 @@
 -- complain if script is sourced in psql, rather than via CREATE EXTENSION
 \echo Use "ALTER EXTENSION h3 UPDATE TO '0.4.0'" to load this file. \quit
 
-CREATE FUNCTION h3_line(h3index, h3index) RETURNS SETOF h3index
+CREATE OR REPLACE FUNCTION h3_line(h3index, h3index) RETURNS SETOF h3index
     AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
     COMMENT ON FUNCTION h3_line(h3index, h3index) IS
 'Given two H3 indexes, return the line of indexes between them (inclusive).