+++ src/inspection.c
@@ -21,8 +21,8 @@
 #include "h3-pg.h"
 
 // Returns the resolution of the index
-PG_FUNCTION_INFO_V1(h3_h3_get_resolution);
-Datum h3_h3_get_resolution(PG_FUNCTION_ARGS)
+PG_FUNCTION_INFO_V1(h3_get_resolution);
+Datum h3_get_resolution(PG_FUNCTION_ARGS)
 {
     H3Index *hex = PG_GETARG_H3_INDEX_P(0);
     int resolution = h3GetResolution(*hex);
@@ -30,8 +30,8 @@ Datum h3_h3_get_resolution(PG_FUNCTION_ARGS)
 }
 
 // Returns the base cell number of the index
-PG_FUNCTION_INFO_V1(h3_h3_get_base_cell);
-Datum h3_h3_get_base_cell(PG_FUNCTION_ARGS)
+PG_FUNCTION_INFO_V1(h3_get_base_cell);
+Datum h3_get_base_cell(PG_FUNCTION_ARGS)
 {
     H3Index *hex = PG_GETARG_H3_INDEX_P(0);
     int base_cell_number = h3GetBaseCell(*hex);
@@ -54,8 +54,8 @@ Datum h3_string_to_h3(PG_FUNCTION_ARGS)
 
 // Converts the internal representation of the index to the string
 // representation
-PG_FUNCTION_INFO_V1(h3_h3_to_string);
-Datum h3_h3_to_string(PG_FUNCTION_ARGS)
+PG_FUNCTION_INFO_V1(h3_to_string);
+Datum h3_to_string(PG_FUNCTION_ARGS)
 {
     H3Index *hex = PG_GETARG_H3_INDEX_P(0);
     char *str = palloc(17 * sizeof(char));
@@ -65,8 +65,8 @@ Datum h3_h3_to_string(PG_FUNCTION_ARGS)
 }
 
 // Returns true if this is a valid H3 index
-PG_FUNCTION_INFO_V1(h3_h3_is_valid);
-Datum h3_h3_is_valid(PG_FUNCTION_ARGS)
+PG_FUNCTION_INFO_V1(h3_is_valid);
+Datum h3_is_valid(PG_FUNCTION_ARGS)
 {
     H3Index *hex = PG_GETARG_H3_INDEX_P(0);
     bool isValid = h3IsValid(*hex);
@@ -74,8 +74,8 @@ Datum h3_h3_is_valid(PG_FUNCTION_ARGS)
 }
 
 // Returns true if this index has a resolution with Class III orientation
-PG_FUNCTION_INFO_V1(h3_h3_is_res_class_iii);
-Datum h3_h3_is_res_class_iii(PG_FUNCTION_ARGS)
+PG_FUNCTION_INFO_V1(h3_is_res_class_iii);
+Datum h3_is_res_class_iii(PG_FUNCTION_ARGS)
 {
     H3Index *hex = PG_GETARG_H3_INDEX_P(0);
     bool isResClassIII = h3IsResClassIII(*hex);
@@ -83,8 +83,8 @@ Datum h3_h3_is_res_class_iii(PG_FUNCTION_ARGS)
 }
 
 // Returns true if this hex represents a pentagonal cell
-PG_FUNCTION_INFO_V1(h3_h3_is_pentagon);
-Datum h3_h3_is_pentagon(PG_FUNCTION_ARGS)
+PG_FUNCTION_INFO_V1(h3_is_pentagon);
+Datum h3_is_pentagon(PG_FUNCTION_ARGS)
 {
     H3Index *hex = PG_GETARG_H3_INDEX_P(0);
     bool isPentagon = h3IsPentagon(*hex);