+++ src/miscellaneous.c
@@ -16,7 +16,10 @@
 
 #include <postgres.h> // Datum, etc.
 #include <fmgr.h>     // PG_FUNCTION_ARGS, etc.
+#include <funcapi.h>  // Definitions for functions which return sets
+
 #include <h3/h3api.h> // Main H3 include
+#include "h3-pg.h"
 
 // Converts degrees to radians
 PG_FUNCTION_INFO_V1(h3_degs_to_rads);
@@ -79,4 +82,26 @@ Datum h3_num_hexagons(PG_FUNCTION_ARGS)
     int resolution = PG_GETARG_INT32(0);
     unsigned long long retVal = numHexagons(resolution);
     PG_RETURN_INT64(retVal);
+}
+
+// Provides all resolution 0 indexes
+PG_FUNCTION_INFO_V1(h3_get_res_0_indexes);
+Datum h3_get_res_0_indexes(PG_FUNCTION_ARGS)
+{
+    if (SRF_IS_FIRSTCALL())
+    {
+        FuncCallContext *funcctx = SRF_FIRSTCALL_INIT();
+        MemoryContext oldcontext =
+            MemoryContextSwitchTo(funcctx->multi_call_memory_ctx);
+
+        int count = res0IndexCount();
+        H3Index *indexes = palloc(sizeof(H3Index) * count);
+        getRes0Indexes(indexes);
+
+        funcctx->user_fctx = indexes;
+        funcctx->max_calls = count;
+        MemoryContextSwitchTo(oldcontext);
+    }
+
+    SRF_RETURN_H3_INDEXES_FROM_USER_FCTX();
 }