+++ src/type.c
@@ -22,28 +22,22 @@
 #include "h3-pg.h"
 
 Datum h3_string_to_h3(PG_FUNCTION_ARGS);
-Datum h3_h3_to_string(PG_FUNCTION_ARGS);
+Datum h3_to_string(PG_FUNCTION_ARGS);
+
 
 PG_FUNCTION_INFO_V1(h3index_in);
-PG_FUNCTION_INFO_V1(h3index_out);
-PG_FUNCTION_INFO_V1(h3index_eq);
-PG_FUNCTION_INFO_V1(h3index_ne);
-PG_FUNCTION_INFO_V1(h3index_lt);
-PG_FUNCTION_INFO_V1(h3index_le);
-PG_FUNCTION_INFO_V1(h3index_gt);
-PG_FUNCTION_INFO_V1(h3index_ge);
-PG_FUNCTION_INFO_V1(h3index_cmp);
-
 Datum h3index_in(PG_FUNCTION_ARGS)
 {
     return h3_string_to_h3(fcinfo);
 }
 
+PG_FUNCTION_INFO_V1(h3index_out);
 Datum h3index_out(PG_FUNCTION_ARGS)
 {
-    return h3_h3_to_string(fcinfo);
+    return h3_to_string(fcinfo);
 }
 
+PG_FUNCTION_INFO_V1(h3index_eq);
 Datum h3index_eq(PG_FUNCTION_ARGS)
 {
     H3Index *a = PG_GETARG_H3_INDEX_P(0);
@@ -51,6 +45,7 @@ Datum h3index_eq(PG_FUNCTION_ARGS)
     PG_RETURN_BOOL(*a == *b);
 }
 
+PG_FUNCTION_INFO_V1(h3index_ne);
 Datum h3index_ne(PG_FUNCTION_ARGS)
 {
     H3Index *a = PG_GETARG_H3_INDEX_P(0);
@@ -58,6 +53,7 @@ Datum h3index_ne(PG_FUNCTION_ARGS)
     PG_RETURN_BOOL(*a != *b);
 }
 
+PG_FUNCTION_INFO_V1(h3index_lt);
 Datum h3index_lt(PG_FUNCTION_ARGS)
 {
     H3Index *a = PG_GETARG_H3_INDEX_P(0);
@@ -65,6 +61,7 @@ Datum h3index_lt(PG_FUNCTION_ARGS)
     PG_RETURN_BOOL(*a < *b);
 }
 
+PG_FUNCTION_INFO_V1(h3index_le);
 Datum h3index_le(PG_FUNCTION_ARGS)
 {
     H3Index *a = PG_GETARG_H3_INDEX_P(0);
@@ -72,6 +69,7 @@ Datum h3index_le(PG_FUNCTION_ARGS)
     PG_RETURN_BOOL(*a <= *b);
 }
 
+PG_FUNCTION_INFO_V1(h3index_gt);
 Datum h3index_gt(PG_FUNCTION_ARGS)
 {
     H3Index *a = PG_GETARG_H3_INDEX_P(0);
@@ -79,6 +77,7 @@ Datum h3index_gt(PG_FUNCTION_ARGS)
     PG_RETURN_BOOL(*a > *b);
 }
 
+PG_FUNCTION_INFO_V1(h3index_ge);
 Datum h3index_ge(PG_FUNCTION_ARGS)
 {
     H3Index *a = PG_GETARG_H3_INDEX_P(0);
@@ -86,6 +85,7 @@ Datum h3index_ge(PG_FUNCTION_ARGS)
     PG_RETURN_BOOL(*a >= *b);
 }
 
+PG_FUNCTION_INFO_V1(h3index_cmp);
 Datum h3index_cmp(PG_FUNCTION_ARGS)
 {
     H3Index *a = PG_GETARG_H3_INDEX_P(0);