+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Renamed `h3_basecells` to `h3_get_res_0_indexes`.
+- Renamed all functions with double `h3_h3_` prefix to use single `h3_` prefix.
 
 ## [0.4.0] - 2019-01-12
 ### Added